requirements.txt
alembic==1.6.2
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==41.0.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
Flask-Migrate==2.1.1
flask-restx==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
manage==0.1.15
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
pycparser==2.20
PyJWT==1.5.3
PyMySQL
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==1.0.1
